+++ src/Events/InvoiceVoided.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Emad\FeeCollection\Events;
+
+use Emad\FeeCollection\Models\AccountStatement;
+
+class InvoiceVoided
+{
+    public function __construct(public AccountStatement $statement) {}
+}